Выносит просулшиватель модальных окон в utils, делает функцию универсально
