+++ js/main.js
@@ -1,5 +1,6 @@
 import {createObjects} from './data.js';
 import {createMiniatures} from './create-miniatures.js';
+import './upload-photo-form.js';
 
 createObjects();
 createMiniatures(createObjects());
